+++ src/quacc/settings.py
@@ -175,17 +175,13 @@ class QuaccSettings(BaseSettings):
     )
 
     # VASP Settings: General
-    VASP_INCAR_COPILOT: bool = Field(
-        True,
+    VASP_INCAR_COPILOT: Literal["off", "on", "aggressive"] = Field(
+        "on",
         description=(
-            "Whether co-pilot mode should be used for VASP INCAR handling."
-            "This will modify INCAR flags on-the-fly if they disobey the VASP manual."
-        ),
-    )
-    VASP_COPILOT_OVERRIDE: bool = Field(
-        False,
-        description=(
-            "Whether to force co-pilot swaps to override user-specified flags."
+            "Controls VASP co-pilot mode for automated INCAR parameter handling."
+            "off: Do not use co-pilot mode. INCAR parameters will be unmodified."
+            "on: Use co-pilot mode. This will only modify INCAR flags not already set by the user."
+            "aggressive: Use co-pilot mode in agressive mode. This will modify INCAR flags even if they are already set by the user."
         ),
     )
     VASP_BADER: bool = Field(